energi/etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["uch <uch@energinet.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cfgrib = "^0.9.14.1"
azure-identity = "^1.18.0"
azure-functions = "^1.21.3"
azure-storage-file-datalake = "^12.17.0"
zarr = "^2.18.3"
fsspec = "^2024.10.0"
adlfs = "^2024.7.0"
pandas = "^2.2.3"
scipy = "^1.14.1"
cftime = "^1.6.4"
xarray = "^2024.10.0"
python-dotenv = "^1.0.1"
dask = "^2024.11.2"
h5netcdf = "^1.6.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"